Move FrilansFinansApi models to their own folder
