+++ lib/frilans_finans_api/lib/frilans_finans_api/models/company.rb
