+++ spec/lib/frilans_finans_api/models/company_spec.rb
