+++ spec/lib/frilans_finans_api/models/curency_spec.rb
