+++ spec/lib/frilans_finans_api/models/invoice_spec.rb
