+++ spec/lib/frilans_finans_api/models/profession_spec.rb
