+++ spec/lib/frilans_finans_api/models/user_spec.rb
